requirements.txt
numpy==1.19.2
tensorflow-cpu
Keras
Flask
gunicorn
Pillow